Added basic shadow rendering
On level load, or when the rebuildShadowMap command is issued, a shadow map will be generating, using up
to a fixed number of lights and the uploaded to the graphicscard. When rendering with lights enabled this
shadow map will then be used, to determin how far each light shines in a vertain direction.
!!IMPORTANT!! If new lights are added in the editor, the shadow map is not automatically rebuild. You
will have to do so yourself, by using the rebuildShadowMap command.
are added in the editor,