Rebuild the Screen Coordinate System to be more useful for gui Design. Added scissoring to the Widget rendering.
The Screen Coordinate System now has its origin in the top left corner, and extends downward and to the right,
more in accordance with classical ui systems. The Height of the coordinate system is currently 20.

The Widget rendering now uses scissoring to prevent widgets from rendering outside of their actual area.
This enables the creation of scrollable areas and similar widgets.